refactor(style-modify): move toolbar related files
